Remove obsolete KillTasks message from MarathonSchedulerActor
Killing tasks via the API is now always directly processed by the KillService. Before, when the `wipe` flag was not specified as `true`, killing tasks was indirected via the MarathonSchedulerService, and silently failed if the associated app was locked by a deployment, even if `force=true` was specified. With this patch, it is possible to kill tasks even if they are subject of a running deployment.

JIRA issues: MARATHON-1686